saved_searches updated to use timeZoneAndFormat from configRecords
