+++ packages/netsuite-adapter/src/filter.ts
@@ -19,6 +19,7 @@ import NetsuiteClient from './client/client'
 import { LazyElementsSourceIndexes } from './elements_source_index/types'
 import { DeployResult } from './types'
 import { NetsuiteConfig } from './config'
+import { TimeZoneAndFormat } from './changes_detector/date_formats'
 
 
 export type Filter = filter.Filter<void, DeployResult>
@@ -28,6 +29,7 @@ export type LocalFilterOpts = {
   elementsSource: ReadOnlyElementsSource
   isPartial: boolean
   config: NetsuiteConfig
+  timeZoneAndFormat?: TimeZoneAndFormat
   changesGroupId?: string
   fetchTime?: Date
 }
